test(functools): add tests for `optree.functools`
